start migrating auth tests to pytest
